+++ jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  // we only run jest in subfolders
+  testMatch: ["none"]
+}